But fix in generating random number
maxRange was multiplied directly to the random
number generated using Math.random(). If the start
range is a non-zero value, then the start value is
added to the random number multipled with maxRange.
It creates a bug if the pre-add value is
maxRange - 1.

Instead of multiplying maxRange, the random number
should be multiplied with (maxRange - minRange).